pyproject.toml
[project]
name = "simple-ai-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-gigachat>=0.3.10",
    "langgraph>=0.4.5",
    "python-dotenv>=1.1.0",
]
